+++ src/neuroglancer/ui/layer_list_panel.ts
@@ -151,7 +151,10 @@ class LayerListItem extends RefCounted {
     const deleteButton = makeDeleteButton({
       title: 'Delete layer',
       onClick: () => {
-        deleteLayer(this.layer);
+        const deleteConfirmText = `Are you sure you want to delete the annotation layer ?\nAnnotation layer name: ${this.layer.name}`;
+        if(confirm(deleteConfirmText)) {
+          deleteLayer(this.layer);
+        }
       }
     });
     deleteButton.classList.add('neuroglancer-layer-list-panel-item-delete');
